# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume writing service API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ServicePackage {
  id: string;
  name: string;
  slug: string;
  price: number;
  deliveryDays: number;
  description: string;
  features: string[];
  isPopular: boolean;
}

export interface AddOn {
  id: string;
  name: string;
  price: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending_payment: "pending_payment",
  pending_verification: "pending_verification",
  confirmed: "confirmed",
  in_progress: "in_progress",
  delivered: "delivered",
} as const;

export interface Order {
  id: string;
  orderId: string;
  name: string;
  email: string;
  phone: string;
  targetJobRole: string;
  workExperience: string;
  /** @nullable */
  additionalNotes?: string | null;
  servicePackage: string;
  addOns: string[];
  totalAmount: number;
  /** @nullable */
  resumeFileUrl?: string | null;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  /** @nullable */
  transactionId?: string | null;
  status: OrderStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateOrderBody {
  name: string;
  email: string;
  phone: string;
  targetJobRole: string;
  workExperience: string;
  /** @nullable */
  additionalNotes?: string | null;
  servicePackage: string;
  addOns: string[];
  totalAmount: number;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending_payment: "pending_payment",
  pending_verification: "pending_verification",
  confirmed: "confirmed",
  in_progress: "in_progress",
  delivered: "delivered",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface UploadPaymentProofBody {
  paymentScreenshotUrl: string;
  /** @nullable */
  transactionId?: string | null;
}

export interface OrderStats {
  total: number;
  pendingPayment: number;
  pendingVerification: number;
  confirmed: number;
  inProgress: number;
  delivered: number;
  totalRevenue: number;
}