# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume writing service API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all available resume service packages with pricing and features
 * @summary List all service packages
 */
export const ListServicesResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  slug: zod.string(),
  price: zod.number(),
  deliveryDays: zod.number(),
  description: zod.string(),
  features: zod.array(zod.string()),
  isPopular: zod.boolean(),
});
export const ListServicesResponse = zod.array(ListServicesResponseItem);

/**
 * Submit a new resume writing order with customer details
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  targetJobRole: zod.string(),
  workExperience: zod.string(),
  additionalNotes: zod.string().nullish(),
  servicePackage: zod.string(),
  addOns: zod.array(zod.string()),
  totalAmount: zod.number(),
});

/**
 * Returns all orders for the admin panel
 * @summary List all orders (admin)
 */
export const ListOrdersResponseItem = zod.object({
  id: zod.string(),
  orderId: zod.string(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  targetJobRole: zod.string(),
  workExperience: zod.string(),
  additionalNotes: zod.string().nullish(),
  servicePackage: zod.string(),
  addOns: zod.array(zod.string()),
  totalAmount: zod.number(),
  resumeFileUrl: zod.string().nullish(),
  paymentScreenshotUrl: zod.string().nullish(),
  transactionId: zod.string().nullish(),
  status: zod.enum([
    "pending_payment",
    "pending_verification",
    "confirmed",
    "in_progress",
    "delivered",
  ]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Get a specific order
 */
export const GetOrderParams = zod.object({
  orderId: zod.coerce.string(),
});

export const GetOrderResponse = zod.object({
  id: zod.string(),
  orderId: zod.string(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  targetJobRole: zod.string(),
  workExperience: zod.string(),
  additionalNotes: zod.string().nullish(),
  servicePackage: zod.string(),
  addOns: zod.array(zod.string()),
  totalAmount: zod.number(),
  resumeFileUrl: zod.string().nullish(),
  paymentScreenshotUrl: zod.string().nullish(),
  transactionId: zod.string().nullish(),
  status: zod.enum([
    "pending_payment",
    "pending_verification",
    "confirmed",
    "in_progress",
    "delivered",
  ]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update order status (admin)
 */
export const UpdateOrderStatusParams = zod.object({
  orderId: zod.coerce.string(),
});

export const UpdateOrderStatusBody = zod.object({
  status: zod.enum([
    "pending_payment",
    "pending_verification",
    "confirmed",
    "in_progress",
    "delivered",
  ]),
});

export const UpdateOrderStatusResponse = zod.object({
  id: zod.string(),
  orderId: zod.string(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  targetJobRole: zod.string(),
  workExperience: zod.string(),
  additionalNotes: zod.string().nullish(),
  servicePackage: zod.string(),
  addOns: zod.array(zod.string()),
  totalAmount: zod.number(),
  resumeFileUrl: zod.string().nullish(),
  paymentScreenshotUrl: zod.string().nullish(),
  transactionId: zod.string().nullish(),
  status: zod.enum([
    "pending_payment",
    "pending_verification",
    "confirmed",
    "in_progress",
    "delivered",
  ]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Upload payment screenshot and transaction ID
 */
export const UploadPaymentProofParams = zod.object({
  orderId: zod.coerce.string(),
});

export const UploadPaymentProofBody = zod.object({
  paymentScreenshotUrl: zod.string(),
  transactionId: zod.string().nullish(),
});

export const UploadPaymentProofResponse = zod.object({
  id: zod.string(),
  orderId: zod.string(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string(),
  targetJobRole: zod.string(),
  workExperience: zod.string(),
  additionalNotes: zod.string().nullish(),
  servicePackage: zod.string(),
  addOns: zod.array(zod.string()),
  totalAmount: zod.number(),
  resumeFileUrl: zod.string().nullish(),
  paymentScreenshotUrl: zod.string().nullish(),
  transactionId: zod.string().nullish(),
  status: zod.enum([
    "pending_payment",
    "pending_verification",
    "confirmed",
    "in_progress",
    "delivered",
  ]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * Summary counts by status for admin dashboard
 * @summary Get order statistics (admin)
 */
export const GetOrderStatsResponse = zod.object({
  total: zod.number(),
  pendingPayment: zod.number(),
  pendingVerification: zod.number(),
  confirmed: zod.number(),
  inProgress: zod.number(),
  delivered: zod.number(),
  totalRevenue: zod.number(),
});