# lib/api-zod/src/generated/types/addOn.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume writing service API
 * OpenAPI spec version: 0.1.0
 */

export interface AddOn {
  id: string;
  name: string;
  price: number;
}