# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume writing service API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  name: string;
  email: string;
  phone: string;
  targetJobRole: string;
  workExperience: string;
  /** @nullable */
  additionalNotes?: string | null;
  servicePackage: string;
  addOns: string[];
  totalAmount: number;
}