# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume writing service API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  orderId: string;
  name: string;
  email: string;
  phone: string;
  targetJobRole: string;
  workExperience: string;
  /** @nullable */
  additionalNotes?: string | null;
  servicePackage: string;
  addOns: string[];
  totalAmount: number;
  /** @nullable */
  resumeFileUrl?: string | null;
  /** @nullable */
  paymentScreenshotUrl?: string | null;
  /** @nullable */
  transactionId?: string | null;
  status: OrderStatus;
  createdAt: string;
  updatedAt: string;
}