# lib/api-zod/src/generated/types/orderStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume writing service API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderStats {
  total: number;
  pendingPayment: number;
  pendingVerification: number;
  confirmed: number;
  inProgress: number;
  delivered: number;
  totalRevenue: number;
}