# lib/api-zod/src/generated/types/servicePackage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume writing service API
 * OpenAPI spec version: 0.1.0
 */

export interface ServicePackage {
  id: string;
  name: string;
  slug: string;
  price: number;
  deliveryDays: number;
  description: string;
  features: string[];
  isPopular: boolean;
}