# lib/api-zod/src/generated/types/uploadPaymentProofBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Resume writing service API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadPaymentProofBody {
  paymentScreenshotUrl: string;
  /** @nullable */
  transactionId?: string | null;
}